Add sidebar with working animations and scrollTo
I use javascript and CSS to animate the sidebar. I had to make a cloned
sidebar component so that it can serve as the 'closing' animation, while
the real one disappears from display. I had to do this because the close
animation wouldn't work unless the element is reloaded into the DOM.

I created empty spans above each component, since the sidebar's links
would scrollTo the component without leaving any space for the header,
obstructing part of the sections. I made the spans absolute and had them
rest 82px above the actual components, since that is the height of the
header.